+++ Libplanet/Store/BaseStore.cs
@@ -3,13 +3,21 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Globalization;
 using System.Linq;
+using System.Numerics;
 using System.Security.Cryptography;
+using Bencodex.Types;
 using Libplanet.Action;
+using Libplanet.Assets;
 using Libplanet.Blocks;
 using Libplanet.Tx;
+using Serilog;
+using FAV = Libplanet.Assets.FungibleAssetValue;
 
 namespace Libplanet.Store
 {
+    /// <summary>
+    /// Common code for several <see cref="IStore"/> implementations.
+    /// </summary>
     public abstract class BaseStore : IStore, IDisposable
     {
         /// <inheritdoc />
@@ -121,6 +129,15 @@ namespace Libplanet.Store
         /// <inheritdoc cref="IStore.ContainsBlock(BlockHash)"/>
         public abstract bool ContainsBlock(BlockHash blockHash);
 
+        /// <inheritdoc cref="IStore.PutTxExecution(Libplanet.Tx.TxSuccess)"/>
+        public abstract void PutTxExecution(TxSuccess txSuccess);
+
+        /// <inheritdoc cref="IStore.PutTxExecution(Libplanet.Tx.TxFailure)"/>
+        public abstract void PutTxExecution(TxFailure txFailure);
+
+        /// <inheritdoc cref="IStore.GetTxExecution(BlockHash, TxId)"/>
+        public abstract TxExecution GetTxExecution(BlockHash blockHash, TxId txid);
+
         /// <inheritdoc cref="IStore.SetBlockPerceivedTime(BlockHash, DateTimeOffset)"/>
         public abstract void SetBlockPerceivedTime(
             BlockHash blockHash,
@@ -160,5 +177,126 @@ namespace Libplanet.Store
 
         /// <inheritdoc/>
         public abstract void ForkTxNonces(Guid sourceChainId, Guid destinationChainId);
+
+        protected static IValue SerializeTxExecution(TxSuccess txSuccess)
+        {
+            var sDelta = new Dictionary(
+                txSuccess.UpdatedStates.Select(kv =>
+                    new KeyValuePair<IKey, IValue>(
+                        new Binary(kv.Key.ByteArray),
+                        kv.Value is { } v ? List.Empty.Add(v) : List.Empty
+                    )
+                )
+            );
+            var favDelta = SerializeGroupedFAVs(txSuccess.FungibleAssetsDelta);
+            var updatedFAVs = SerializeGroupedFAVs(txSuccess.UpdatedFungibleAssets);
+            return (Dictionary)Dictionary.Empty
+                .Add("fail", false)
+                .Add("sDelta", sDelta)
+                .Add((IKey)(Text)"favDelta", new Dictionary(favDelta))
+                .Add((IKey)(Text)"updatedFAVs", new Dictionary(updatedFAVs));
+        }
+
+        protected static IValue SerializeTxExecution(TxFailure txFailure)
+        {
+            Dictionary d = Dictionary.Empty
+                .Add("fail", true)
+                .Add("exc", txFailure.ExceptionName);
+            return txFailure.ExceptionMetadata is { } v ? d.Add("excMeta", v) : d;
+        }
+
+        protected static TxExecution DeserializeTxExecution(
+            BlockHash blockHash,
+            TxId txid,
+            IValue decoded,
+            ILogger logger
+        )
+        {
+            if (!(decoded is Bencodex.Types.Dictionary d))
+            {
+                const string msg = nameof(TxExecution) +
+                    " must be serialized as a Bencodex dictionary, not {ActualValue}.";
+                logger?.Error(msg, decoded.Inspection);
+                return null;
+            }
+
+            try
+            {
+                bool fail = d.GetValue<Bencodex.Types.Boolean>("fail");
+                if (fail)
+                {
+                    string excName = d.GetValue<Text>("exc");
+                    IValue excMetadata = d.ContainsKey("excMeta") ? d["excMeta"] : null;
+                    return new TxFailure(blockHash, txid, excName, excMetadata);
+                }
+
+                ImmutableDictionary<Address, IValue> sDelta = d.GetValue<Dictionary>("sDelta")
+                    .ToImmutableDictionary(
+                        kv => new Address((Binary)kv.Key),
+                        kv => kv.Value is List l && l.Any() ? l[0] : null
+                    );
+                IImmutableDictionary<Address, IImmutableDictionary<Currency, FungibleAssetValue>>
+                    favDelta = DeserializeGroupedFAVs(d.GetValue<Dictionary>("favDelta"));
+                IImmutableDictionary<Address, IImmutableDictionary<Currency, FungibleAssetValue>>
+                    updatedFAVs = DeserializeGroupedFAVs(d.GetValue<Dictionary>("updatedFAVs"));
+                return new TxSuccess(
+                    blockHash,
+                    txid,
+                    sDelta,
+                    favDelta,
+                    updatedFAVs
+                );
+            }
+            catch (Exception e)
+            {
+                const string msg =
+                    "Uncaught exception during deserializing a " + nameof(TxExecution) +
+                    ": {Exception}";
+                logger?.Error(e, msg, e);
+                return null;
+            }
+        }
+
+        private static Bencodex.Types.Dictionary SerializeGroupedFAVs(
+            IImmutableDictionary<Address, IImmutableDictionary<Currency, FAV>> balanceDelta
+        ) =>
+            new Dictionary(
+                balanceDelta.Select(pair =>
+                    new KeyValuePair<IKey, IValue>(
+                        new Binary(pair.Key.ByteArray),
+                        SerializeFAVs(pair.Value)
+                    )
+                )
+            );
+
+        private static IImmutableDictionary<Address, IImmutableDictionary<Currency, FAV>>
+        DeserializeGroupedFAVs(Bencodex.Types.Dictionary serialized) =>
+            serialized.ToImmutableDictionary(
+                kv => new Address((Binary)kv.Key),
+                kv => DeserializeFAVs((List)kv.Value)
+            );
+
+        private static Bencodex.Types.List SerializeFAVs(
+            IImmutableDictionary<Currency, FungibleAssetValue> favs
+        ) =>
+            new List(
+                favs.Select(
+                    kv => (IValue)List.Empty.Add(kv.Key.Serialize()).Add(kv.Value.RawValue)
+                )
+            );
+
+        private static IImmutableDictionary<Currency, FungibleAssetValue> DeserializeFAVs(
+            List serialized
+        ) =>
+            serialized.Select(pList =>
+            {
+                var pair = (List)pList;
+                var currency = new Currency(pair[0]);
+                BigInteger rawValue = (Bencodex.Types.Integer)pair[1];
+                return new KeyValuePair<Currency, FAV>(
+                    currency,
+                    FAV.FromRawValue(currency, rawValue)
+                );
+            }).ToImmutableDictionary();
     }
 }